tech(quest-claim): move confirm modal to component directory
